Code cleanups assited by the IntelliJ IDEA code analyzer.
git-svn-id: https://svn.codehaus.org/izpack/izpack-src/trunk@1925 7d736ef5-cfd4-0310-9c9a-b52d5c14b761

+++ src/lib/com/izforge/izpack/gui/TwoColumnLayout.java
@@ -760,7 +760,7 @@ public class TwoColumnLayout implements LayoutManager2
     private int topBuffer(int usedHeight, Container parent)
     {
         int amount = ((int) parent.getSize().getHeight()) - usedHeight;
-        amount = (int) (amount * topBuffer) / 100;
+        amount = amount * topBuffer / 100;
 
         return (amount);
     }
